Fixes filenames in WAF
CKAN requires that filenames must end with .xml in any WAF it attempts
to harvest from. Unfortunately nginx's autoindex function truncates
longer filenames.

This patch truncates all filenames to 43 characters and appends .xml to
the end so that CKAN will pick it. So filenames MUST be unique up to 43
characters.

It also removes any query params which look bad anyway.