Applies a hash for filenames on central WAF
Before, the registry would only use the first 43 characters of a
filename. If we hash the full file name, we reduce the risk of collision
to almost 0 and we still key uniqueness based on the file name. If two
filenames from two different harvests are the same, a collision will
occur which is still the desired behavior.