Fixes "infinite" loop in CSW harvest
Removes an "infinite" (actually limited by max_batches parameter) loop in the
CSW harvesting code.  nextRecord is set to 0 when all the records have
been fetched, but the code was checking if nextRecord-1 was equal to the
number of matched records, which appeared to be unsatisfiable.  Thus,
redundant fetches were made until the maximum batches were exceeded.